+++ R/zzz.R
@@ -44,5 +44,5 @@ opt.orgdata <- list(
 }
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("orgdata version 0.4.3")
+  packageStartupMessage("orgdata version 0.4.4")
 }